chore: improve Radio exports
